Servo control

+++ Mitsubishi/01_ServoControl/DataTransferObject/PLCConnector.cs
@@ -42,7 +42,7 @@ namespace _01_ServoControl.BusinessLogicLayer
         /// <param name="device">Địa chỉ trên PLC</param>
         /// <param name="data">Dữ liệu truyền xuống PLC</param>
         /// <returns>Return true nếu set thành công</returns>
-        public static bool SetDevice(string device, int data)
+        public static bool SetDevice(string device, short data)
         {
             return MXDriver.SetDevice(device, data);
         }
@@ -52,7 +52,7 @@ namespace _01_ServoControl.BusinessLogicLayer
         /// <param name="device">Địa chỉ cần lấy data</param>
         /// <param name="data">Data trả về</param>
         /// <returns>Return true nếu lấy data thành công</returns>
-        public static bool GetDevice(string device, out int data)
+        public static bool GetDevice(string device, out short data)
         {
             return MXDriver.GetDevice(device, out data);
         }
